Move mobile-cards. Fix styles
