+++ src/components/mobile-cards/index.js
@@ -2,6 +2,6 @@ import { connect } from 'react-redux';
 
 import * as actions from 'actions/url-actions';
 
-import Cards from './cards-component';
+import Cards from './mobile-cards-component';
 
 export default connect(null, actions)(Cards);
