+++ src/containers/scenes/mobile/nrc-landing-scene-mobile/nrc-landing-scene-mobile-component.jsx
@@ -7,10 +7,10 @@ import { useT } from '@transifex/react';
 import CountriesBordersLayer from 'containers/layers/countries-borders-layer';
 import CountryLabelsLayer from 'containers/layers/country-labels-layer';
 import ArcgisLayerManager from 'containers/managers/arcgis-layer-manager';
-import Cards from 'containers/mobile/nrc-landing/cards';
 import { useOnboardingTooltipRefs } from 'containers/onboarding/onboarding-hooks';
 
 import CountryEntryTooltip from 'components/country-entry-tooltip';
+import Cards from 'components/mobile-cards';
 import Scene from 'components/scene';
 import SearchLocation from 'components/search-location';
 
@@ -89,7 +89,7 @@ function NrcLandingMobileComponent({
           <BackArrowIcon className={styles.arrowIcon} />
         </button>
 
-        <p>{t('National reports cards')}</p>
+        <p>{t('National report cards')}</p>
       </header>
 
       <div className={styles.searchContainer}>
